+++ src/vector_intergrator.rs
@@ -1,28 +1,28 @@
 use crate::*;
 
 pub trait VecIntegrator{
-    fn integrate_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f32> ;
-    fn integrate_uniform_parallel(&self, integral_method : &impl VecIntegralMethod, funct : & FunVector) -> Vec<f32> ;
-    fn integrate_non_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f32> ;
-    fn integrate_non_uniform_parallel(&self, integral_method : &impl VecIntegralMethod, funct : & FunVector) -> Vec<f32> ;
+    fn integrate_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f64> ;
+    fn integrate_uniform_parallel(&self, integral_method : &impl VecIntegralMethod, funct : & FunVector) -> Vec<f64> ;
+    fn integrate_non_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f64> ;
+    fn integrate_non_uniform_parallel(&self, integral_method : &impl VecIntegralMethod, funct : & FunVector) -> Vec<f64> ;
     fn number_of_points_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> i32;
     fn number_of_points_uniform_parallel(&self, integral_method : &impl VecIntegralMethod, funct : & FunVector) -> i32;
     fn number_of_points_non_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<i32>;
     fn number_of_points_non_uniform_parallel(&self, integral_method : &impl VecIntegralMethod, funct : & FunVector) -> Vec<i32>;
-    fn relative_errors_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f32> {
+    fn relative_errors_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f64> {
         integral_method.relative_error_uniform(funct, self.number_of_points_uniform(integral_method, funct))
     }
-    fn relative_errors_non_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f32> {
+    fn relative_errors_non_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f64> {
         integral_method.relative_error_non_uniform(funct,self.number_of_points_non_uniform(integral_method,funct))
     }
 }
 
 pub struct VecFixedPrecision {
-    pub precision : f32 ,
+    pub precision : f64 ,
 }
 
 impl VecIntegrator for VecFixedPrecision {
-    fn integrate_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f32>  {
+    fn integrate_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f64>  {
         let mut number_of_points: i32 = 0;
         let dimension = funct.components.len();
         let mut condition : Vec<bool> = vec![false;dimension];
@@ -53,7 +53,7 @@ impl VecIntegrator for VecFixedPrecision {
         integral_method.integrate_uniform(funct, number_of_points)
     }
 
-    fn integrate_uniform_parallel(&self, integral_method : &impl VecIntegralMethod, funct : & FunVector) -> Vec<f32>  {
+    fn integrate_uniform_parallel(&self, integral_method : &impl VecIntegralMethod, funct : & FunVector) -> Vec<f64>  {
         let mut number_of_points: i32 = 0;
         let dimension = funct.components.len();
         let mut condition : Vec<bool> = vec![false;dimension];
@@ -148,7 +148,7 @@ impl VecIntegrator for VecFixedPrecision {
         number_of_points
     }
 
-    fn integrate_non_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f32>  {
+    fn integrate_non_uniform(&self, integral_method : &impl VecIntegralMethod, funct : & FunctVector) -> Vec<f64>  {
         let mut number_of_points: i32 = 0;
         let dimension = funct.components.len();
         let mut number_of_points_vec : Vec<i32> = vec![1;dimension];
@@ -188,7 +188,7 @@ impl VecIntegrator for VecFixedPrecision {
         integral_method.integrate_non_uniform(funct, number_of_points_vec)
     }
 
-    fn integrate_non_uniform_parallel(&self, integral_method : &impl VecIntegralMethod, funct : & FunVector) -> Vec<f32>  {
+    fn integrate_non_uniform_parallel(&self, integral_method : &impl VecIntegralMethod, funct : & FunVector) -> Vec<f64>  {
         let mut number_of_points: i32 = 0;
         let dimension = funct.components.len();
         let mut number_of_points_vec : Vec<i32> = vec![1;dimension];
